+++ container-engine-lib/lib/docker_manager/create_and_start_container_args.go
@@ -5,14 +5,6 @@ import (
 	"net"
 )
 
-type PortPublishOption string
-
-const (
-	NoPortPUblishing
-)
-
-var noManualPortPublishings = map[nat.Port]nat.PortBinding{}
-
 // See CreateAndStartContainerArgsBuilder for detailed documentation on the fields
 type CreateAndStartContainerArgs struct {
 	dockerImage                    string
@@ -23,9 +15,7 @@ type CreateAndStartContainerArgs struct {
 	staticIp                       net.IP
 	addedCapabilities              map[ContainerCapability]bool
 	networkMode                    DockerManagerNetworkMode
-	usedPortsSet                   map[nat.Port]bool
-	shouldAutoPublishAllPorts      bool	   // Incompatible with manual port publishing
-	manualPortPublishings          map[nat.Port]nat.PortBinding // Incompatible with auto port publishing
+	usedPorts                      map[nat.Port]PortPublishSpec
 	entrypointArgs                 []string
 	cmdArgs                        []string
 	envVariables                   map[string]string
@@ -45,9 +35,7 @@ type CreateAndStartContainerArgsBuilder struct {
 	staticIp                       net.IP
 	addedCapabilities              map[ContainerCapability]bool
 	networkMode                    DockerManagerNetworkMode
-	usedPortsSet                   map[nat.Port]bool
-	shouldAutoPublishAllPorts      bool	   // Incompatible with manual port publishing
-	manualPortPublishings          map[nat.Port]nat.PortBinding // Incompatible with auto port publishing
+	usedPorts                      map[nat.Port]PortPublishSpec
 	entrypointArgs                 []string
 	cmdArgs                        []string
 	envVariables                   map[string]string
@@ -73,9 +61,7 @@ func NewCreateAndStartContainerArgsBuilder(dockerImage string, name string, netw
 		staticIp:                       nil,
 		addedCapabilities:              map[ContainerCapability]bool{},
 		networkMode:                    DefaultNetworkMode,
-		usedPortsSet:                   map[nat.Port]bool{},
-		shouldAutoPublishAllPorts:      false,
-		manualPortPublishings:          noManualPortPublishings,
+		usedPorts:                      map[nat.Port]PortPublishSpec{},
 		entrypointArgs:                 nil,
 		cmdArgs:                        nil,
 		envVariables:                   map[string]string{},
@@ -97,8 +83,7 @@ func (builder *CreateAndStartContainerArgsBuilder) Build() *CreateAndStartContai
 		staticIp:                       builder.staticIp,
 		addedCapabilities:              builder.addedCapabilities,
 		networkMode:                    builder.networkMode,
-		usedPortsSet:                   builder.usedPortsSet,
-		shouldAutoPublishAllPorts:      builder.shouldAutoPublishAllPorts,
+		usedPorts:                      builder.usedPorts,
 		entrypointArgs:                 builder.entrypointArgs,
 		cmdArgs:                        builder.cmdArgs,
 		envVariables:                   builder.envVariables,
@@ -140,23 +125,11 @@ func (builder *CreateAndStartContainerArgsBuilder) WithNetworkMode(mode DockerMa
 	return builder
 }
 
-// A set of ports that the container will listen on
-func (builder *CreateAndStartContainerArgsBuilder) WithUsedPorts(portsSet map[nat.Port]bool) *CreateAndStartContainerArgsBuilder {
-	builder.usedPortsSet = portsSet
+// A set of ports that the container will listen on, mapped to a specification for how they should be published to the host machine (if at all)
+func (builder *CreateAndStartContainerArgsBuilder) WithUsedPorts(usedPorts map[nat.Port]PortPublishSpec) *CreateAndStartContainerArgsBuilder {
+	builder.usedPorts = usedPorts
 	return builder
 }
-
-// If true, we'll publish all the exposed ports to the Docker host on ephemeral ports so that the outside world can connect
-//  to the container
-func (builder *CreateAndStartContainerArgsBuilder) ShouldAutoPublishAllPorts(shouldAutoPublishAllPorts bool) *CreateAndStartContainerArgsBuilder {
-	builder.shouldAutoPublishAllPorts = shouldAutoPublishAllPorts
-	builder.with
-	return builder
-}
-
-// Defines a manual mapping of container ports -> ports on the host machine
-// If set, will disable ShouldAutoPublishAllPorts
-func (builder *CreateAndStartContainerArgsBuilder) WithManualPortPublishings
 
 // The args that will be used to override the ENTRYPOINT of the image (leave as nil to not override)
 func (builder *CreateAndStartContainerArgsBuilder) WithEntrypointArgs(args []string) *CreateAndStartContainerArgsBuilder {