+++ container-engine-lib/docs/changelog.md
@@ -2,6 +2,11 @@
 ### Changes
 * Clarified that the `all` argument to `GetContainersByLabels` is for whether stopped containers should be shown
 
+### Breaking Changes
+* The `CreateAndStartContainerArgsBuilder.WithUsedPorts`'s parameter now has a value of `PortPublishSpec`, which defines how the port should be published
+* `CreateAndStartContainerArgsBuilder.ShouldPublishAllPorts` parameter has been removed
+    * Users should migrate to `CreateAndStartContainerArgsBuilder.WithUsedPorts` instead
+
 # 0.6.0
 ### Features
 * Add `Network` type to store Docker Network information as ID, name, ip and mask